beacon-serve/requirements.txt
flask==0.10
cassandra-driver>=1.0.2
